Add availability metric for mirror
This allows us to monitor the response status code for a URL, which can
use to guage whether the mirror is available. A better measure for
availability might be using synthetics to request a wider range of URLs
from the mirror and using error rate to determine availability. However
this is probably go enough for now.